# backend/openapi.yaml
openapi: 3.1.0
info:
  title: ctf-blockchain-infra
  version: "1.0.0"
  description: |
    HTTP surface of the ctf-blockchain-infra backend. The public
    endpoints (`/api/{config,status,flag,sign,launch,kill,reset,instance,rpc,health}`)
    are unauthenticated and per-IP rate-limited. The admin surface
    (`/api/admin/*`) requires a bearer token.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://ctf.example.com
    description: Default deployment URL pattern
  - url: http://127.0.0.1:8787
    description: Local dev backend
...
security: []
paths:
...
  /api/writeup/{id}:
    post:
      tags: [public]
      summary: Submit a writeup. Gated by on-chain isSolved(player). Optionally signed.
      parameters:
        - $ref: "#/components/parameters/ChallengeId"
        - $ref: "#/components/parameters/AddressQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [writeup]
              properties:
                writeup:   { type: string, maxLength: 4096 }
                timestamp: { type: integer, description: "ms since epoch — required when signing" }
                signature: { type: string,  description: "personal_sign over the canonical writeup message — required when signing" }
      responses:
        "200":
          description: Accepted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted: { type: boolean }
                  ts:       { type: string }
                  bytes:    { type: integer }
                  signed:   { type: boolean }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: signature does not match address.
        "403":
          description: player has not satisfied isSolved yet.
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /api/status/{id}:
    get:
      tags: [public]
      summary: Has this player satisfied isSolved on this challenge?
      parameters:
        - $ref: "#/components/parameters/ChallengeId"
        - $ref: "#/components/parameters/AddressQuery"
      responses:
        "200":
          description: Status snapshot
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Status" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "502": { $ref: "#/components/responses/RpcError" }
  /api/flag/{id}:
    get:
      tags: [public]
      summary: Read the flag if the player has satisfied isSolved.
      parameters:
        - $ref: "#/components/parameters/ChallengeId"
        - $ref: "#/components/parameters/AddressQuery"
      responses:
        "200":
          description: Solve result; flag included only when solved is true.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Flag" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/sign/{id}:
    get:
      tags: [public]
      summary: Backend-issued signature (only on challenges with signer.enabled).
      parameters:
        - $ref: "#/components/parameters/ChallengeId"
        - $ref: "#/components/parameters/AddressQuery"
      responses:
        "200":
          description: Either personal-sign or eip712-typed-data signature.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/PersonalSignResult"
                  - $ref: "#/components/schemas/Eip712SignResult"
        "400": { $ref: "#/components/responses/BadRequest" }
        "404":
          description: Challenge has no signer configured.
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/launch/{id}:
    post:
      tags: [launcher]
      summary: Spawn a per-player anvil container.
      parameters:
        - $ref: "#/components/parameters/ChallengeId"
        - $ref: "#/components/parameters/AddressQuery"
      responses:
        "200":
          description: Instance metadata.
          content:
            application/json:
              schema:
                type: object
                properties:
                  instance: { $ref: "#/components/schemas/Instance" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/ServerError" }
        "503":
          description: Launcher disabled on this server.
  /api/kill/{id}:
    post:
      tags: [launcher]
      summary: Stop and remove the player's container.
      parameters:
        - $ref: "#/components/parameters/ChallengeId"
        - $ref: "#/components/parameters/AddressQuery"
      responses:
        "200":
          description: "`{ killed: bool }`"
          content:
            application/json:
              schema:
                type: object
                properties:
                  killed: { type: boolean }
  /api/reset/{id}:
    post:
      tags: [launcher]
      summary: Revert the player's anvil state to immediately after deploy.
      parameters:
        - $ref: "#/components/parameters/ChallengeId"
        - $ref: "#/components/parameters/AddressQuery"
      responses:
        "200":
          description: Reset succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  reset: { type: boolean }
        "400":
          description: No instance, no snapshot, or wrong mode.
  /api/instance/{id}:
    get:
      tags: [launcher]
      summary: Read the player's current instance, or null.
      parameters:
        - $ref: "#/components/parameters/ChallengeId"
        - $ref: "#/components/parameters/AddressQuery"
      responses:
        "200":
          description: Instance or null.
          content:
            application/json:
              schema:
                type: object
                properties:
                  instance:
                    oneOf:
                      - $ref: "#/components/schemas/Instance"
                      - type: "null"
  /api/rpc/{instanceId}:
    post:
      tags: [launcher]
      summary: |
        Method-allowlisted JSON-RPC proxy to the player's anvil
        container. Allowed prefixes: `eth_`, `net_`, `web3_`, `anvil_`,
        `evm_`, `debug_`. Denied: `personal_*`, `miner_*`, `admin_*`,
        `txpool_*`.
      parameters:
        - in: path
          name: instanceId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: JSON-RPC 2.0 request, single or batch.
              oneOf:
                - type: object
                - type: array
                  items: { type: object }
      responses:
        "200": { description: JSON-RPC response (or batch). }
        "400": { description: invalid request. }
        "403": { description: method not in allowlist. }
        "404": { description: unknown instance. }
        "410": { description: instance expired. }
        "429": { $ref: "#/components/responses/RateLimited" }
        "502": { description: upstream anvil error. }
...
components:
  securitySchemes:
    AdminBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    ChallengeId:
      in: path
      name: id
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9_-]{1,32}$"
    AddressQuery:
      in: query
      name: address
      required: true
      schema:
        type: string
        format: address
        example: "0xAbC1234567890abcDEF1234567890ABCdef123456"
  responses:
    BadRequest:
      description: Malformed input.
      content:
        application/json:
          schema:
            type: object
            properties: { error: { type: string } }
    RateLimited:
      description: Rate limited.
      content:
        application/json:
          schema:
            type: object
            properties: { error: { type: string } }
    RpcError:
      description: Upstream RPC failure.
      content:
        application/json:
          schema:
            type: object
            properties: { error: { type: string } }
    ServerError:
      description: Unexpected server error.
  schemas:
...
    Status:
      type: object
      properties:
        solved:  { type: boolean }
        spawned: { type: boolean }
        target:
          oneOf: [{ type: string }, { type: "null" }]
        instance:
          oneOf:
            - $ref: "#/components/schemas/Instance"
            - type: "null"
    Flag:
      type: object
      properties:
        solved: { type: boolean }
        flag:
          type: string
          description: Present only when solved is true.
    PersonalSignResult:
      type: object
      properties:
        type:      { const: personal-sign }
        player:    { type: string }
        signer:    { type: string }
        signature: { type: string }
    Eip712SignResult:
      type: object
      properties:
        type:      { const: eip712 }
        player:    { type: string }
        signer:    { type: string }
        signature: { type: string }
        primaryType: { type: string, nullable: true }
        domain:    { type: object }
        types:     { type: object }
        message:   { type: object }
    Instance:
      type: object
      properties:
        instanceId:     { type: string }
        fullInstanceId: { type: string }
        mode:           { type: string, enum: [proxy, direct] }
        rpcUrl:         { type: string }
        target:         { type: string }
        extra:
          type: object
          additionalProperties: true
        createdAt: { type: integer }
        expiresAt: { type: integer }